Utilsクラスの作成 #20
